+++ __mocks__/blocklypy.js
@@ -1,0 +1,5 @@
+module.exports = {
+    convertProjectToPython: jest.fn(),
+    IPyConverterFile: {},
+    IPyConverterOptions: {},
+};