+++ src/commands/connect-device.ts
@@ -1,17 +1,28 @@
 import * as vscode from 'vscode';
-import { TreeCommands } from '../extension/tree-commands';
 import { Device } from '../logic/ble';
 import { StateProp, withState } from '../logic/state';
 
 const items: vscode.QuickPickItem[] = [];
 
+export async function connectDeviceAsyncAny(...args: any[]): Promise<any> {
+    await connectDeviceAsync(args[0]);
+}
+
 export async function connectDeviceAsync(name: string) {
     if (name.length === 0) throw new Error('No device name provided to connect to.');
 
-    await withState(StateProp.Connected, async () => {
-        await Device.disconnectAsync();
-    })();
+    await vscode.window.withProgress(
+        {
+            location: { viewId: 'blocklypy-vscode-commands' },
+            cancellable: false,
+        },
+        async () => {
+            await withState(StateProp.Connected, async () => {
+                await Device.disconnectAsync();
+            })();
 
-    // if a name is provided, connect directly
-    await Device.connectAsync(name, TreeCommands.refresh.bind(TreeCommands));
+            // if a name is provided, connect directly
+            await Device.connectAsync(name);
+        },
+    );
 }