+++ src/commands/disconnect-device.ts
@@ -1,9 +1,15 @@
 import { TreeCommands } from '../extension/tree-commands';
 import { Device } from '../logic/ble';
+import { hasState, StateProp } from '../logic/state';
+import { stopUserProgramAsync } from './stop-user-program';
 
 export async function disconnectDeviceAsync() {
     if (!Device.current) {
         throw new Error('No device is currently connected.');
+    }
+
+    if (hasState(StateProp.Running)) {
+        await stopUserProgramAsync();
     }
 
     await Device.disconnectAsync();